// apps/console/src/__generated__/core/MeasureEvidencesTabFragment.graphql.ts
export type MeasureEvidencesTabFragment$data = {
  readonly canUploadEvidence: boolean;
  readonly evidences: {
    readonly __id: string;
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly file: {
          readonly fileName: string;
          readonly mimeType: string;
          readonly size: number;
        } | null | undefined;
        readonly id: string;
        readonly " $fragmentSpreads": FragmentRefs<"MeasureEvidencesTabFragment_evidence">;
      };
    }>;
  };
  readonly id: string;
  readonly name: string;
  readonly " $fragmentType": "MeasureEvidencesTabFragment";
};
export type MeasureEvidencesTabFragment$key = {
  readonly " $data"?: MeasureEvidencesTabFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"MeasureEvidencesTabFragment">;
};